refactor(formatter): centralize raw shell syntax helpers
