+++ crates/shuck-formatter/src/word/mod.rs
@@ -7,12 +7,13 @@ use crate::facts::{
     classify_sequence_contains_multiline_literal_source,
 };
 use crate::options::{IndentStyle, ResolvedShellFormatOptions};
-use crate::scan::{
-    QuoteState, RawShellScanner, common_nonempty_shell_indent, heredoc_start,
-    leading_shell_indent as line_leading_shell_indent,
-    line_indent_before_offset as line_indent_before_source_offset,
-    line_without_continuation_backslash, redirect_operator_end, refine_common_indent,
+use crate::raw_syntax::{
+    QuoteState, RawShellScanner, RawShellText, common_nonempty_shell_indent, heredoc_start,
+    leading_shell_indent as line_leading_shell_indent, line_ends_with_raw_continuation_operator,
+    line_without_continuation_backslash, matching_raw_command_substitution_close,
+    normalize_raw_pipeline_continuations, redirect_operator_end, refine_common_indent,
 };
+use crate::scan::line_indent_before_offset as line_indent_before_source_offset;
 use crate::streaming::format_stmt_sequence_streaming_to_buf;
 use shuck_ast::{
     ArithmeticAssignOp, ArithmeticBinaryOp, ArithmeticExpansionSyntax, ArithmeticExpr,
@@ -39,6 +40,5 @@ pub(crate) use self::core::{
 pub(crate) use self::heredoc::render_heredoc_body_to_buf;
 pub(crate) use self::parameter::{parameter_defaulting_operator, render_pattern_syntax_to_buf};
 pub(crate) use self::raw_rewrites::{
-    matching_raw_command_substitution_close, normalize_raw_empty_parameter_replacement_delimiters,
-    normalize_raw_pipeline_continuations, normalize_raw_unquoted_word_continuations,
+    normalize_raw_empty_parameter_replacement_delimiters, normalize_raw_unquoted_word_continuations,
 };
